update Readme to reflect the slight change in goals for Netify
Where originally my plan was to use the project to compile just my notes
into a website. I am now hoping that I can use the structure that I've
built up here to achieve that goal in a separate project.  In that way I
will be able to reuse the Netify starting point for other projects in
the future.