app.py: split NetifyApp into a Core and some Mixins
The goal is to make it easier to use netify as a library to compose your
own custom web apps. This way you can include the core functionality of
Netify and the features that you need. At the moment that just means you
can choose to have a CLI interface, a UWSGI interface or both. In the
future that could mean including a SQLAlchemy back end, Security
features such as login support and other modular pieces.